requirements.txt
yt-dlp
google-api-python-client==2.125.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
gspread==5.7.2
requests==2.31.0